Crate chip8_vm

Virtual machine for the CHIP-8 programming language

This crate implements a virtual machine for the CHIP-8 programming language. It can be used as a backend for CHIP-8 emulators, debuggers and so on.

The code is split into the instructions module, which provides the translation from raw bits (RawInstruction) into valid instructions (Instruction).

The vm module contains the actual virtual machine implementation (Vm).

The error module contains the Chip8Error implementation of std:error::Error for any kinds of errors that might occur using the chip8_vm crate.

Reexports

pub use instructions::*;

Modules

error
std:error:Error implementations

instructions
Raw and high-level instruction abstractions

vm
Virtual machine implementation

Functions

version
Returns the version of this crate in the format MAJOR.MINOR.PATCH.