Struct chip8_vm::instructions::RawInstruction

pub struct RawInstruction {
    // some fields omitted
}

Raw instruction

Helper around the raw bits, not necessarily a valid instruction.

Methods

impl RawInstruction

fn new(bits: u16) -> RawInstruction

Creates a new raw instruction without any checks of the bits

fn bits(&self) -> u16

The raw bits

fn addr(&self) -> Addr

The address bits part

fn x(&self) -> Vx

The Vx register-index part, i.e. 0xE is VE

fn y(&self) -> Vy

The Vy register-index part, i.e. 0xE is VE

fn n_high(&self) -> Nibble

The high nibble part

fn n_low(&self) -> Nibble

The low nibble part

fn k(&self) -> Byte

The k byte part