Struct chip8_vm::vm::Vm

pub struct Vm {
    // some fields omitted
}

Virtual machine

A virtual machine manages state like its registers, the RAM, stack, screen pixels, pressed keys as well as timers and some internal state.

Methods

impl Vm

fn new() -> Vm

fn load_rom(&mut self, reader: &mut Reader) -> Result<usize, Chip8Error>

fn dump_ram(&self, writer: &mut Writer)

fn beeping(&self) -> bool

fn set_key(&mut self, idx: u8)

fn unset_key(&mut self, idx: u8)

fn step(&mut self, dt: f32)

fn screen_rows<'a>(&'a self) -> Chunks<'a, u8>

fn print_screen(&self)

fn print_disassembly(&self)